requirements.txt
keyboard==0.13.5
openai==0.27.6
pyperclip==1.8.2
python-dotenv==1.0.0
